+++ crates/sdk/tests/anchor_test/mod.rs
@@ -24,3 +24,5 @@ mod oracle;
 mod token;
 
 mod callback;
+
+mod competition;