+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - programs(store): Added `GtBuyback` CPI event to be emitted on GT exchange vault confirmation.
 - programs(store): Added `confirm_gt_exchange_vault_v2` instruction, which requires the caller to provide buyback information.
 - programs(store): Added `OrderUpdated` CPI event to be emitted on order creation or update.
+- programs: Added the `gmsol-competition` program.
 - sdk(sdk): Added `callback` option to `ops::CreateOrderBuilder`.
 - sdk(sdk): Added `callback` field to `ExecuteOrderHint` and `CloseOrderHint`.
 
@@ -68,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - programs: Implemented `Default` for `Glv` and made the `store` field public.
 - programs: Introduced a separate `impl_fixed_map!` macro that implements fixed map functionality without defining the corresponding struct.
 - programs: Introduced a separated `impl_flags!` macro that implements flag map functionality without defining the container.
+- programs: Added the `gmsol-competition` program.
 - model: Re-exported `num_traits`.
 - sdk: Added more functions to `SquadsOps`:
   - `SquadsOps::squads_create_vault_transaction_and_return_data`: Creates a vault transaction and return the data.